requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
asgiref==3.5.0
asttokens==2.4.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
distro==1.9.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.2
fasteners==0.14.1
fonttools==4.51.0
future==0.18.2
greenlet==1.1.2
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jeepney==0.7.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kazam==1.4.5
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==1.23.6
packaging==24.0
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
powerline-status==2.8.2
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==3.1.2
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.2
reportlab==3.6.8
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.37.2
systemd-python==234
threadpoolctl==3.4.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
uvicorn==0.29.0
wadllib==1.3.6
wcwidth==0.2.13
wsproto==1.0.0
xdg==5
xgboost==2.0.3
xkit==0.0.0
zipp==1.0.0
